update : code splitting page on /index, /article/[slug], /user/posts
